backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-core==0.1.52
faiss-cpu>=1.7.4
python-dotenv==1.0.0
pypdf==3.17.4
tiktoken>=0.7.0
numpy>=1.26.0
